Stage detection seems to work!
Tested on Lincs 300, Moelhills and Welsh Lakes 200. All stops
were detected correctly, but I do need to now double check
all the calculations to check my assumptions are still valid,
because previously I assumed that the TrackPoint at the end
of a stage was used as the start of the next stage.