Remove command line option for writing the TrackPoints sheet.
We know always produce this sheet, though hyperlinking each
track point to its Google Map location remains optional.

The reason is that the main 'Summary' sheet is to have its
columns reduced - instead of duplicating lat/lon etc. for
points such as max speed and heart rate, we will simply
link to the corresponding point on the Track Points sheet.