Convert to in-memory XML parser.
The BufRead parser cannot read the text inside the Extensions tag.
Only the quick_xml::Reader which is based on a byte slice can do that, so we
have to read the whole file into memory first.

This version still has the 'buf' variables though. They can probably be
eliminated as well.